server/requirements.txt
Flask==3.0.2
Flask-Cors==5.0.0
groq==0.11.0
pymongo==4.10.1
python-dotenv==1.0.1
